+++ test/reduce-two-e2e/reduce_two_test.go
@@ -87,53 +87,6 @@ func (r *ReduceSuite) testReduceStream(lang string) {
 	done <- struct{}{}
 }
 
-func (r *ReduceSuite) TestSimpleSessionPipeline() {
-
-	// the reduce feature is not supported with redis ISBSVC
-	if strings.ToUpper(os.Getenv("ISBSVC")) == "REDIS" {
-		r.T().SkipNow()
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	defer cancel()
-	w := r.Given().Pipeline("@testdata/session-reduce/simple-session-sum-pipeline.yaml").
-		When().
-		CreatePipelineAndWait()
-	defer w.DeletePipelineAndWait()
-	pipelineName := "simple-session-sum"
-
-	// wait for all the pods to come up
-	w.Expect().VertexPodsRunning()
-
-	count := 0
-	done := make(chan struct{})
-	go func() {
-		// publish messages to source vertex, with event time starting from 60000
-		startTime := 60000
-		for i := 0; true; i++ {
-			select {
-			case <-ctx.Done():
-				return
-			case <-done:
-				return
-			default:
-				if count == 10 {
-					startTime = startTime + (10 * 1000)
-					count = 0
-				} else {
-					startTime = startTime + 1000
-				}
-				eventTime := strconv.Itoa(startTime)
-				w.SendMessageTo(pipelineName, "in", NewHttpPostRequest().WithBody([]byte("100")).WithHeader("X-Numaflow-Event-Time", eventTime))
-				count += 1
-			}
-		}
-	}()
-
-	w.Expect().RedisSinkContains("simple-session-sum-sink", "1000")
-	done <- struct{}{}
-}
-
 func (r *ReduceSuite) TestSimpleSessionKeyedPipelineGo() {
 	r.testSimpleSessionKeyedPipeline("go")
 }
@@ -143,7 +96,6 @@ func (r *ReduceSuite) TestSimpleSessionKeyedPipelineJava() {
 }
 
 func (r *ReduceSuite) testSimpleSessionKeyedPipeline(lang string) {
-
 	// the reduce feature is not supported with redis ISBSVC
 	if strings.ToUpper(os.Getenv("ISBSVC")) == "REDIS" {
 		r.T().SkipNow()
@@ -195,7 +147,6 @@ func (r *ReduceSuite) testSimpleSessionKeyedPipeline(lang string) {
 }
 
 func (r *ReduceSuite) TestSimpleSessionPipelineFailOverUsingWAL() {
-
 	// the reduce feature is not supported with redis ISBSVC
 	if strings.ToUpper(os.Getenv("ISBSVC")) == "REDIS" {
 		r.T().SkipNow()
