--- test/reduce-two-e2e/testdata/session-reduce/simple-session-sum-pipeline.yaml
@@ -1,47 +1,0 @@
-apiVersion: numaflow.numaproj.io/v1alpha1
-kind: Pipeline
-metadata:
-  name: simple-session-sum
-spec:
-  watermark:
-    maxDelay: 30s
-  vertices:
-    - name: in
-      scale:
-        min: 1
-      source:
-        http: {}
-    - name: compute-sum
-      partitions: 1
-      udf:
-        container:
-          # https://github.com/numaproj/numaflow-go/tree/main/pkg/sessionreducer/examples/sum
-          image: quay.io/numaio/numaflow-go/session-sum:stable
-          imagePullPolicy: Always
-        groupBy:
-          window:
-            session:
-              timeout: 10s
-          keyed: true
-          storage:
-            persistentVolumeClaim:
-              volumeSize: 2Gi
-              accessMode: ReadWriteOnce
-    - name: sink
-      scale:
-        min: 1
-      sink:
-        udsink:
-          container:
-            # A redis sink for e2e testing, see https://github.com/numaproj/numaflow-go/tree/main/pkg/sinker/examples/redis-sink
-            image: quay.io/numaio/numaflow-go/redis-sink:stable
-            imagePullPolicy: Always
-            env:
-              - name: SINK_HASH_KEY
-                # The key is set in the format of "pipeline_name-vertex_name"
-                value: "simple-session-sum-sink"
-  edges:
-    - from: in
-      to: compute-sum
-    - from: compute-sum
-      to: sink
